+++ CHANGELOG.md
@@ -3,6 +3,28 @@ List of major changes and improvements
 
 ## [Unreleased]
 
+### Added
+
+- Hardware models.
+
+  - Access forwarding.
+
+  - Buffer sharing scheme.
+    - Use `BufShrScheme` class to represent and calculate NoC transfers.
+
+
+### Changed
+
+- Software models.
+
+  - `LoopBlockingScheme` supports access forwarding and buffer sharing.
+
+  - `partition` unit number of hops calculation supports access forwarding and
+    buffer sharing.
+
+  - `DataLayout` supports closest-first forwarding data transfer for access
+    forwarding and buffer sharing.
+
 
 ## [v1.5 -- v1.6] -- 2018-01-31